# Improve performance of spatial queries