@aschampion 2017-01-22
Benchmark plots attached.

On the upside, I don't see any notable edit performance regressions except for skeleton import, which is expected and acceptable. However, I also don't get the expected query performance benefits. Postgres 9.6/PostGIS 2.3. Will look into it more tomorrow.

![new_node_index_query](https://cloud.githubusercontent.com/assets/460206/22184563/108a313c-e0a2-11e6-8a0a-73f6fe471d6f.png)
![new_node_index_benchmarks](https://cloud.githubusercontent.com/assets/460206/22184565/181d3b74-e0a2-11e6-8a38-a6a98b9279ae.png)
![new_node_index_skeleton_benchmarks](https://cloud.githubusercontent.com/assets/460206/22184564/132350ea-e0a2-11e6-8faf-254ff5021d49.png)