@aschampion 2017-01-22
Nevermind, these benchmarks are tainted because my PG tuning got lost in the 16.04 upgrade earlier this week (hence the poor baseline performance on 2017.01.19). I'll rebench tomorrow after tuning.