@Remi-C 2017-01-23
Hey, two slightly obvious ideas, but you never know :
 - you may want to use prepared statements (you do for insert, I don't know
for selects).
 - Also when bulk inserting, it is faster to drop the index and rebuild it
after insert

Cheers

2017-01-22 19:02 GMT+01:00 Andrew Champion <notifications@github.com>:

> Nevermind, these benchmarks are tainted because my PG tuning got lost in
> the 16.04 upgrade earlier this week (hence the poor baseline performance on
> 2017.01.19). I'll rebench tomorrow after tuning.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/catmaid/CATMAID/pull/1481#issuecomment-274347215>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AEgcIsjkAQB8DTuH0Ycp6Ep8EMq4bzCUks5rU5m7gaJpZM4LpzbR>
> .
>