@tomka 2017-01-23
Thanks for running the benchmarks @aschampion and I am curious how it looks after tuning. If it is similar to the current results, I'd be interested in the query plans. The claim of an improvement by an order of magnitude was maybe too bold indeed. While I found it to be true for many queries, some where also only three times faster. I'll do some more benchmarking today as well.

> you may want to use prepared statements (you do for insert, I don't know
for selects).

Thanks for the suggestions! However, my understanding of planned queries is that they mainly remove planning time and live for one session, but please correct me if I am wrong here. While the spatial queries in this issue are executed often, their planning time seems to be around 1ms at the moment. Letting our Django back-end re-create planned queries every time the session is renewed seems like potential overhead if no connection pooling is used (which one of course wants to have in general). Fixing this would require additional code without much gain, but I certainly try it to see what difference it actually makes.

>  Also when bulk inserting, it is faster to drop the index and rebuild it
after insert

Thanks, good to know!