@tomka 2017-01-27
Just for reference: I added the `NODE_PROVIDER` option (to be used in `settings.py`), which can be set to `'classic'`, `'postgis2d'` or `'postgis3d'`. Currently, `'postgis3d'` is the default, because the performance improvements of this branch are still somewhat inconclusive.