@tomka 2017-01-31
I did some more tests that also include orthogonal views and tested how things go with only the `numrange` index and no `2D GIST` on XY, as this seemed to work better with lower node counts in the result. I used seven queries with different node counts and sampled each one (only) three times, but the error seems pretty low. I'll produce some graphs with higher node counts tough. This is what I get on the Neurocity server for complete `node/list` requests (on local loopback interface) with Postgres 9.6.1 and PostGIS 2.3.1:

![node_indices_xy](https://cloud.githubusercontent.com/assets/112226/22472833/c7a4e372-e7a4-11e6-9b93-0a4fdc709cbb.png)

![node_indices_zy](https://cloud.githubusercontent.com/assets/112226/22472837/ca1c1f08-e7a4-11e6-94d1-105078798321.png)

![node_indices_zy2](https://cloud.githubusercontent.com/assets/112226/22472845/cce43ac2-e7a4-11e6-868a-b8949e5f69d4.png)


Like we already noticed, my 3D query times seem to be much higher (absolute and relative to 2D) than what you get @aschampion, so I wonder if you could provide the query plans for a representative query using the current 3D index and one with the 2D+range indices if you find time for it (using the `auto_explain` module seems convenient here).

The graphs also suggest that this index update might help a little bit with orthogonal views or at least it maintains performance. If we would really run into index update timing problems, we could think about dropping either the `2D GIST` index or the `3D GIST`. If we did the former, orthogonal views would need to use the `3D GIST` and would still be usable (and maybe even faster for larger field of views). The index update impact doesn't seem to be a problem at the moment though, so dropping any index is not required at the moment.

I confirmed the relative timing difference on both Neurocity and Neuropil and could see a query speedup of at least factor 2. The numbers initially claimed referred only to the 
spatial query alone (without joins and connectors) and even for this were not completely robust. Anyhow, I feel it would still be nice to merge this, since our production systems seem to benefit from it. It is easy to change the query type (and activate prepared statements) in `settings.py`. So what do you think?