@aschampion 2017-01-31
![new_node_index_query](https://cloud.githubusercontent.com/assets/460206/22478783/5465d0e6-e7b9-11e6-9942-74c65bc9f531.png)

Unfortunately the only performance delta I'm seeing is from the change to the half-open interval (shown by the trend lines). But, definitely good performance changes on other systems and the commits look good, so let's merge.