@tomka 2017-01-31
Thanks for testing again, let's see if use in production will yield more insight in what is different between setups.